pyproject.toml
[tool.poetry]
name = "dsg"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^1.26.2"
pandas = "^2.1.3"
scipy = "^1.11.4"
crcmod = "^1.7"
pyserial = "^3.5"
pillow = "^10.1.0"
fonttools = "^4.46.0"
pyyaml = "^6.0.1"
matplotlib = "^3.8.2"
opencv-python = "^4.8.1.78"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.27.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
